Extracted ProducerPerformance to separate component
