+++ packages/page-staking/src/Performance/BlockProductionCommitteeList.tsx
@@ -10,7 +10,7 @@ import { useLenientThresholdPercentage, useNextTick } from '@polkadot/react-hook
 
 import Filtering from '../Filtering.js';
 import { useTranslation } from '../translate.js';
-import Address from './Address/index.js';
+import ProducerPerformance from '../react-components/ProducerPerformance/index.js';
 
 interface Props {
   className?: string;
@@ -97,7 +97,7 @@ function BlockProductionCommitteeList ({ className, eraValidatorPerformances, ex
       header={headerRef.current}
     >
       {list.map(({ isCommittee, validatorPerformance }): React.ReactNode => (
-        <Address
+        <ProducerPerformance
           address={validatorPerformance.accountId}
           blocksCreated={validatorPerformance.blockCount}
           filterName={nameFilter}
