+++ packages/page-staking/src/Performance/FinalityCommittee.tsx
@@ -7,11 +7,11 @@ import { CardSummary, SummaryBox, Table } from '@polkadot/react-components';
 import { useNextTick } from '@polkadot/react-hooks';
 
 import Filtering from '../Filtering.js';
-import FinalizerAddress from '../react-components/FinalizerPerformance/FinalizerPerformance.js';
 import { useTranslation } from '../translate.js';
 import Legend from './Legend.js';
 import useAbftScores from './useAbftScores.js';
 import { useFinalityCommittee } from './useFinalityCommittee.js';
+import FinalizerPerformance from "../react-components/FinalizerPerformance/index.js";
 
 interface Props {
   className?: string;
@@ -48,7 +48,6 @@ function FinalityCommittee ({ className, currentSession, session }: Props) {
   const headerRef: [string, string, number?][] =
     [
       [t('finalizers'), 'start', 1],
-      [t('session'), 'expand'],
       [t('ABFT score'), 'expand'],
       [t('stats'), 'expand']
     ];
@@ -93,13 +92,12 @@ function FinalityCommittee ({ className, currentSession, session }: Props) {
         legend={<Legend />}
       >
         {list.map(({ abftScore, accountId }): React.ReactNode => (
-          <FinalizerAddress
+          <FinalizerPerformance
             abftScore={abftScore}
             address={accountId}
             filterName={nameFilter}
             key={accountId}
             scoresEnabled={scoresEnabled}
-            session={session}
           />
         ))}
       </Table>
